💄 add header and toolbar to login
