+++ front-end/src/validation/CreateBadgeValidation.tsx
@@ -5,12 +5,3 @@ export const createBadgeSchema = yup.object().shape({
     badgeDescription:yup.string().required(),
     badgeChallenge: yup.string().required(),
 })
-
-export const createUerSchema = yup.object().shape({
-    email: yup.string().required(),
-    name: yup.string().required(),
-    surname:yup.string().required(),
-    number: yup.string().required(),
-    username: yup.string().required(),
-    password:yup.string().required(),
-})
